resources/scripts/dremio-data-generator/requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Faker==3.0.0
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-core==1.1.0
google-cloud-spanner==1.13.0
google-cloud-firestore==1.6.0
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
grpcio-gcp==0.2.2
idna==2.8
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
text-unidecode==1.3
urllib3==1.25.7
